VolunteerNetwork/requirements.txt
django==2.2
djangorestframework
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
